remove "warning: comparison between signed and unsigned integer expressions"
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@4147 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ source/texk/web2c/luatexdir/lua/llfslibext.c
@@ -81,7 +81,7 @@ static int Preadlink(lua_State * L)
 /** readlink(path) */
     const char *path = luaL_checkstring(L, 1);
     char *b = NULL;
-    size_t allocated = 128;
+    int allocated = 128;
     int n;
     while (1) {
         b = malloc(allocated);
